Switch to Kysely + Postgres (#7)
It turns out that SQLite doesn't work well over NFS, which is what AWS
EFS uses under the hood. Specifically, corruption can occur due to
proper locking not being supported over NFS.

Switch to Postgres as the backend, and while we're at it switch to
Kysely as the SQL query generator. This matches what we're doing in our
existing backend, so it will be easier for the team to contribute.